requirements.txt
django
psycopg2_binary==2.9.9
djangorestframework
django-cors-headers
